Refactor: MainSystem Player Modify key binding
