+++ Assets/Scripts/Runtime/CH1/Main/Player/TopDownPlayer.cs
@@ -1,7 +1,5 @@
 using Runtime.CH1.Main.PlayerFunction;
-using Runtime.Data.Original;
 using Runtime.ETC;
-using Runtime.InGameSystem;
 using Runtime.Interface;
 using UnityEngine;
 using UnityEngine.InputSystem;
@@ -23,26 +21,12 @@ namespace Runtime.CH1.Main.Player
         
         private PlayerState _state = PlayerState.Idle;
         private Vector2 _movementInput;
-        private IProvider<ControlsData> ControlsDataProvider => DataProviderManager.Instance.ControlsDataProvider;
-
+        
         private void Start()
         {
             _movement = new TopDownMovement(moveSpeed, transform);
             _animation = new TopDownAnimation(GetComponent<Animator>(), animSpeed);
             _interaction = new TopDownInteraction(transform, LayerMask.GetMask(GlobalConst.Interaction));
-            
-            PlayerKeyBinding();
-        }
-        
-        private void PlayerKeyBinding()
-        {
-            GameOverControls gameControls = ControlsDataProvider.Get().GameOverControls;
-            
-            gameControls.Player.Enable();
-            gameControls.Player.Move.performed += OnMove;
-            gameControls.Player.Move.started += OnMove;
-            gameControls.Player.Move.canceled += OnMove;
-            gameControls.Player.Interaction.performed += ctx => OnInteraction();
         }
         
         private void Update()
@@ -56,12 +40,12 @@ namespace Runtime.CH1.Main.Player
             _state = isMove ? PlayerState.Move : PlayerState.Idle;
         }
         
-        private void OnInteraction()
+        public void OnInteraction()
         {
             bool isInteract = _interaction.Interact(_movement.Direction);
             _state = isInteract ? PlayerState.Interact : PlayerState.Idle;
         }
         
-        private void OnMove(InputAction.CallbackContext context) => _movementInput = context.ReadValue<Vector2>();
+        public void OnMove(InputAction.CallbackContext context) => _movementInput = context.ReadValue<Vector2>();
     }
 }
